requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.0.0
numpy==1.24.0

# Database
supabase==2.3.4
python-dotenv==1.0.0

# ML & NLP
scikit-learn==1.3.0
spacy==3.7.2
nltk==3.8.1

# Document Processing
pypdf==3.17.4
python-docx==1.1.0

# UI Components
streamlit-tags==1.2.8
plotly==5.18.0

# Others
Pillow==10.2.0
